Adds the rating field as an additional Rappor output in city_randomizer.
The possible values for ratings are the integers from 0 through 10
inclusive. We use basic Rappor (no Bloom filters) and represent the
rating as a bit string of length 11 with a 1 in the position
corresponding to the rating--in other words as the number 2^(rating)

+++ utils/data.py
@@ -51,6 +51,6 @@ def readEntries(file_name):
   """
   with file_util.openForReading(file_name) as csvfile:
     reader = csv.reader(csvfile)
-    return [Entry(int(row[0]), row[1], row[2], int(row[3]), float(row[4]),
+    return [Entry(int(row[0]), row[1], row[2], int(row[3]), int(row[4]),
                   row[5])
         for row in reader]
